+++ src/cli/update/action.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, Copy, Clone, Default)]
+pub enum Action {
+    #[default]
+    Update,
+    CheckEnabled,
+}